WIP: Working cnv selectors for top mutated genes menu
