+++ client/dom/GeneSetEdit/radioWithContent.ts
@@ -31,8 +31,7 @@ export function makeRadiosWithContentDivs(options: any, div: Elem) {
 				.style('padding-left', '25px')
 				.style('display', 'block')
 
-			//If no content (i.e. callback), continue
-			if (d.callback) await d.callback(contentDiv)
+			await d.callback(contentDiv)
 			d3select(this).property('checked', true)
 			inputs.property('disabled', false)
 		})
